+++ client/src/components/Footer/index.js
@@ -1,9 +1,7 @@
 import './styles.footer.css';
 
-const Footer = () => {
+export const Footer = () => {
   return (
     <footer>BYE BYE</footer>
   );
 };
-
-export default Footer;
