+++ client/src/pages/LoginPage/styles.login.css
@@ -1,5 +1,5 @@
 .App {
-  background-image: url("./assets/Mask.svg");
+  background-image: url("../../assets/Mask.svg");
   background-position: right;
   min-height: 100vh;
   width: 100%;
